requirements.txt
Django == 1.9.5
paramiko == 2.0.0
python_crontab == 2.0.2
